# artifacts/api-server/python/routes/export.py
from __future__ import annotations
import uuid
import csv
import io
import os
from pathlib import Path
from fastapi import APIRouter, HTTPException
from fastapi.responses import FileResponse

from models.schemas import ExportInput, SqlExportInput, ExportResult
from services.duckdb_service import EXPORTS_DIR

router = APIRouter()


def _clean_exports(keep_latest: int = 50):
    files = sorted(EXPORTS_DIR.glob("*"), key=lambda f: f.stat().st_mtime, reverse=True)
    for f in files[keep_latest:]:
        try:
            os.unlink(f)
        except Exception:
            pass


@router.post("/export/csv")
async def export_csv(body: ExportInput):
    _clean_exports()
    file_name = (body.fileName or "export").rstrip(".csv") + ".csv"
    file_id = str(uuid.uuid4())
    file_path = EXPORTS_DIR / f"{file_id}_{file_name}"

    output = io.StringIO()
    writer = csv.DictWriter(output, fieldnames=body.columns)
    writer.writeheader()
    for row in body.rows:
        safe_row = {k: str(v) if v is not None else "" for k, v in row.items() if k in body.columns}
        writer.writerow(safe_row)

    content = output.getvalue().encode("utf-8")
    with open(file_path, "wb") as f:
        f.write(content)

    return ExportResult(
        downloadUrl=f"/api/export/download/{file_id}_{file_name}",
        fileName=file_name,
        sizeBytes=len(content)
    ).model_dump()


@router.post("/export/excel")
async def export_excel(body: ExportInput):
    try:
        import openpyxl
    except ImportError:
        raise HTTPException(status_code=500, detail="openpyxl not available")

    _clean_exports()
    file_name = (body.fileName or "export").rstrip(".xlsx") + ".xlsx"
    file_id = str(uuid.uuid4())
    file_path = EXPORTS_DIR / f"{file_id}_{file_name}"

    wb = openpyxl.Workbook()
    ws = wb.active
    ws.title = "Query Results"
    ws.append(body.columns)
    for row in body.rows:
        ws.append([row.get(col, "") for col in body.columns])

    wb.save(file_path)
    size = os.path.getsize(file_path)

    return ExportResult(
        downloadUrl=f"/api/export/download/{file_id}_{file_name}",
        fileName=file_name,
        sizeBytes=size
    ).model_dump()


@router.post("/export/sql")
async def export_sql(body: SqlExportInput):
    _clean_exports()
    file_name = (body.fileName or "query").rstrip(".sql") + ".sql"
    file_id = str(uuid.uuid4())
    file_path = EXPORTS_DIR / f"{file_id}_{file_name}"

    content = f"-- Exported SQL Query\n-- Generated by AI SQL Generator\n\n{body.sql}\n"
    encoded = content.encode("utf-8")
    with open(file_path, "wb") as f:
        f.write(encoded)

    return ExportResult(
        downloadUrl=f"/api/export/download/{file_id}_{file_name}",
        fileName=file_name,
        sizeBytes=len(encoded)
    ).model_dump()


@router.get("/export/download/{file_name}")
async def download_export(file_name: str):
    file_path = EXPORTS_DIR / file_name
    if not file_path.exists():
        raise HTTPException(status_code=404, detail="Export file not found")
    safe_name = "_".join(file_name.split("_")[1:]) if "_" in file_name else file_name
    return FileResponse(
        path=str(file_path),
        filename=safe_name,
        media_type="application/octet-stream"
    )