# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI SQL Generator & Data Analytics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /datasets:
    get:
      operationId: listDatasets
      tags: [datasets]
      summary: List all uploaded datasets
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dataset"
    post:
      operationId: uploadDataset
      tags: [datasets]
      summary: Upload a CSV or XLSX dataset (multipart/form-data, handled by client directly)
      responses:
        "201":
          description: Dataset uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
        "400":
          description: Invalid file
  /datasets/{datasetId}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: Get dataset details
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dataset details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
        "404":
          description: Not found
    delete:
      operationId: deleteDataset
      tags: [datasets]
      summary: Delete a dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /datasets/{datasetId}/rows:
    get:
      operationId: getDatasetRows
      tags: [datasets]
      summary: Preview first 100 rows of a dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Preview rows and columns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetPreview"
  /datasets/{datasetId}/columns:
    get:
      operationId: getDatasetColumns
      tags: [datasets]
      summary: Get column schema for a dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Column schema
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ColumnInfo"
  /connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List database connections
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DbConnection"
    post:
      operationId: createConnection
      tags: [connections]
      summary: Add a database connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DbConnectionInput"
      responses:
        "201":
          description: Connection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbConnection"
  /connections/{connectionId}:
    delete:
      operationId: deleteConnection
      tags: [connections]
      summary: Delete a database connection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /connections/{connectionId}/test:
    post:
      operationId: testConnection
      tags: [connections]
      summary: Test a database connection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
...
  /history:
    get:
      operationId: listHistory
      tags: [history]
      summary: Get query history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: conversationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryEntry"
  /history/{historyId}:
    get:
      operationId: getHistoryEntry
      tags: [history]
      summary: Get a specific history entry
      parameters:
        - name: historyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: History entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"
    delete:
      operationId: deleteHistoryEntry
      tags: [history]
      summary: Delete a history entry
      parameters:
        - name: historyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /export/csv:
    post:
      operationId: exportCsv
      tags: [export]
      summary: Export query results as CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: CSV download info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"
  /export/excel:
    post:
      operationId: exportExcel
      tags: [export]
      summary: Export query results as Excel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: Excel download info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"
  /export/sql:
    post:
      operationId: exportSql
      tags: [export]
      summary: Export as SQL file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SqlExportInput"
      responses:
        "200":
          description: SQL download info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"
...
components:
  schemas:
...
    Dataset:
      type: object
      required: [id, name, fileName, fileType, rowCount, columnCount, createdAt, tableName]
      properties:
        id:
          type: string
        name:
          type: string
        fileName:
          type: string
        fileType:
          type: string
          enum: [csv, xlsx]
        rowCount:
          type: integer
        columnCount:
          type: integer
        sizeBytes:
          type: ["integer", "null"]
        createdAt:
          type: string
        tableName:
          type: string
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnInfo"
    DatasetPreview:
      type: object
      required: [columns, rows, totalRows]
      properties:
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnInfo"
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        totalRows:
          type: integer
    ColumnInfo:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        nullable:
          type: boolean
        sampleValues:
          type: array
          items: {}
    DbConnection:
      type: object
      required: [id, name, dbType, createdAt, isActive]
      properties:
        id:
          type: string
        name:
          type: string
        dbType:
          type: string
          enum: [duckdb, mysql, postgresql]
        host:
          type: ["string", "null"]
        port:
          type: ["integer", "null"]
        database:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        createdAt:
          type: string
        isActive:
          type: boolean
    DbConnectionInput:
      type: object
      required: [name, dbType]
      properties:
        name:
          type: string
        dbType:
          type: string
          enum: [duckdb, mysql, postgresql]
        host:
          type: ["string", "null"]
        port:
          type: ["integer", "null"]
        database:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        connectionString:
          type: ["string", "null"]
    ConnectionTestResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    HistoryEntry:
      type: object
      required: [id, question, sql, createdAt]
      properties:
        id:
          type: string
        question:
          type: string
        sql:
          type: string
        explanation:
          type: ["string", "null"]
        datasetId:
          type: ["string", "null"]
        datasetName:
          type: ["string", "null"]
        connectionId:
          type: ["string", "null"]
        conversationId:
          type: ["string", "null"]
        rowCount:
          type: ["integer", "null"]
        executionTimeMs:
          type: ["number", "null"]
        provider:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ExportInput:
      type: object
      required: [rows, columns]
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        columns:
          type: array
          items:
            type: string
        fileName:
          type: ["string", "null"]
    SqlExportInput:
      type: object
      required: [sql]
      properties:
        sql:
          type: string
        fileName:
          type: ["string", "null"]
    ExportResult:
      type: object
      required: [downloadUrl, fileName]
      properties:
        downloadUrl:
          type: string
        fileName:
          type: string
        sizeBytes:
          type: ["integer", "null"]
...