# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI SQL Generator & Data Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: string;
  question: string;
  sql: string;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  datasetId?: string | null;
  /** @nullable */
  datasetName?: string | null;
  /** @nullable */
  connectionId?: string | null;
  /** @nullable */
  conversationId?: string | null;
  /** @nullable */
  rowCount?: number | null;
  /** @nullable */
  executionTimeMs?: number | null;
  /** @nullable */
  provider?: string | null;
  createdAt: string;
}